% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BasisVectors.R
\name{BasisVectors}
\alias{BasisVectors}
\title{Generate Basis Vectors based on number of enzymes}
\usage{
BasisVectors(NumberOfEnzymes)
}
\arguments{
\item{NumberOfEnzymes}{the number of enzymes used in pathway}
}
\value{
NormalizedVectors
}
\description{
This function takes number of enzymes as input and
defines the orthonormal basis for the hyperplane perpendicular
to the vector <1,1,1...1> in Rn, where n is equal to the number
of enzymes
}
