% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DotProduct.R
\name{DotProduct}
\alias{DotProduct}
\title{Defining dot product}
\usage{
DotProduct(x1, x2)
}
\arguments{
\item{x1}{a vector - explain}

\item{x2}{a vector - explain}
}
\value{
dot product to be used to generate basis vectors
}
\description{
Defining dot product
}
