% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EnzymeModel.R
\name{EnzymeModel}
\alias{EnzymeModel}
\title{EnzymeModel}
\usage{
EnzymeModel(
  NumberOfEnzymes,
  InitialState,
  Time,
  dt,
  EnzymeNumberOfSubstrates,
  EnzymeSubstrates,
  EnzymeProducts,
  EnzymeConcentrations,
  EnzymeTurnoverRates,
  EnzymeMichaelisConstants,
  EnzymeS1CompetitiveInhibitors,
  EnzymeS1CompetitiveInhibitionConstants,
  EnzymeS2CompetitiveInhibitors,
  EnzymeS2CompetitiveInhibitionConstants,
  EnzymeNonCompetitiveInhibitors,
  EnzymeNonCompetitiveInhibitionConstants,
  FinalProduct
)
}
\arguments{
\item{NumberOfEnzymes}{is a scalar value equal to the number of enzymes in the system}

\item{InitialState}{a vector containing the initial concentrations of each fo the species involved in the system. The last entry in InitialState should be a zero, and won't represent any species in the system. This last entry being zero is used for indexing purposes by the function.}
}
\description{
Function that constructs a model for the optimization of enzyme ratios
}
