% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EnzymeModelTimeCourse.R
\name{EnzymeModelTimeCourse}
\alias{EnzymeModelTimeCourse}
\title{EnzymeModelTimeCourse}
\usage{
EnzymeModelTimeCourse(
  NumberOfEnzymes,
  InitialState,
  Time,
  dt,
  EnzymeNumberOfSubstrates,
  EnzymeSubstrates,
  EnzymeProducts,
  EnzymeConcentrations,
  EnzymeTurnoverRates,
  EnzymeMichaelisConstants,
  EnzymeS1CompetitiveInhibitors,
  EnzymeS1CompetitiveInhibitionConstants,
  EnzymeS2CompetitiveInhibitors,
  EnzymeS2CompetitiveInhibitionConstants,
  EnzymeNonCompetitiveInhibitors,
  EnzymeNonCompetitiveInhibitionConstants,
  FinalProduct
)
}
\arguments{
\item{NumberOfEnzymes}{is a scalar value equal to the number of enzymes in the system}

\item{InitialState}{is a vector containing the initial concentrations of each fo the species involved in the system. The last entry in InitialState should be a zero, and won't represent any species in the system. This last entry being zero is used for indexing purposes by the function.}

\item{Time}{is a scalar value that represents how long the simulated time course should represent. The units for time will depend on the units of the inputs to the function. For example, if turnover rates of enzymes are expressed in inverse seconds, the units of time will be seconds.}

\item{dt}{is a scalar value that represents the time step used in the integration of the differential equations representing the enzyme time course. As dt decreases, the computational cost increases, but accuracy of the model also increases.}

\item{EnzymeNumberOfSubstrates}{is a vector of size n, where n is equal to the number of enzymes. For example, if the first enzyme is a single substrate enzyme, then the first entry in EnzymeNumberOfSubstrates should be 1. If the second enzyme in the pathway is a bisubstrate enzyme, then the second entry in EnzymeNumberOfSubstrates should be 2. It is important to note that these functions do not support more than two substrates per enzyme.}

\item{EnzymeSubstrates}{is a matrix of size 2xn, where n is equal to the number of enzymes. Its entries are which species in the InitialState vector are substrates to the nth enzyme in the pathway. For example, if the first enzyme has one substrate and it is compound 1 listed in InitialState, then the first column of the EnzymeSubstrates matrix is 1,0. There must be two entries into the matrix for each column, and if one of the slots isnt used then the value should be 0. The first entry in the column shoul d be filled first. If the second enzyme has two substrates and they are compounds 2 and 3, then the second column of EnzymeSubstrates should be 2,3.}

\item{EnzymeProducts}{is also a matrix of size 2xn, with the same structure and properties as EnzymeSubstrates. However, the entries should represent the products of the enzyme rather than the substrates.}

\item{EnzymeConcentrations}{is a vector of size n. The first entry in the vector should be the concentration of the first enzyme in the pathway. The second entry is the concentration of the second enzyme, and so on.}

\item{EnzymeTurnoverRates}{is a vector of size n. It should be filled in the same manner as EnzymeConcentrations.}

\item{EnzymeMichaelisConstants}{is a matrix of dimensions 2xn. The entries in the matrix are the corresponding Michaelis constants for the enzymatic catalysis of each of the substrates. For example, in our example for EnzymeSubstrates, we said that the first column of EnzymeSubstrates was 1,0. The corresopnding column of EnzymeMichaelisConstants should be Km,1, where Km is the Michaelis constant for the catalysis of compound 1 by enzyme 1 in this example. If a Michaelis Constant does not exist because an enzyme is single substrate, the second entry in the column should be a 1. However, any nonzero constant will work here.}

\item{EnzymeS1CompetitiveInhibitors}{is a matrix of size 10xn, where n is the number of enzymes in the system. Each column of the matrix contains the competitive inhibitors with the first substrate listed in the corresponding EnzymeSubstrates matrix. If there are unused slots, they should be filled in with the final entry of InitialState (Assuming that the final entry is the entry that doesn't represent a species and is just zero). For example, if InitialState is a vector of size 12, then unused slots in the columns of EnzymeS1CompetitiveInhibitors should have the value 12. If compounds 5,6,7 are competitive inhibitors of compound 1 for enzyme 1, then the first column of the matrix is <5,6,7,12,12,12,12,12,12,12>, where 12 is the terminal entry of InitialState.}

\item{EnzymeS1CompetitiveInhibitionConstants}{is a matrix of the same dimensions as EnzymeS1CompetitiveInhibitors, and contains the corresponding competitive inhibition constants for each of the competitive inhibitors with the first substrate binding site. The competitive inhibition constants associated with unused competitive inhibitors should be arbitrary nonzero constants, it is recommended to use zero.}

\item{EnzymeS2CompetitiveInhibitors}{is structurally the same as its S1 counterpart, but the inputs differ based on whether the compounds are competitive inhibitors with the first substrate or the second substrate listed in the EnzymeSubstrates matrix.}

\item{EnzymeS2CompetitiveInhibitionConstants}{is structurally the same as its S2 counterpart, but the inputs differ based on whether the compounds are competitive inhibitors with the first substrate or the second substrate listed in the EnzymeSubstrates matrix.}

\item{EnzymeNonCompetitiveInhibitors}{is also a matrix of size 10xn and should be filled in the same manner as the competitive inhibition matrices.}

\item{FinalProduct}{ask pat}

\item{EnzymeNonCompetitiveInhibitorConstants}{is also a matrix of size 10xn and should be filled in the same manner as the competitive inhibition matrices.}
}
\value{
TimeCourseMatrix
}
\description{
This function outputs the time course for each species as a row in the matrix. The InitialState vector is extended for n entries in each row based on the time step and dt chosen.
}
