% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GradientDescent.R
\name{GradientDescent}
\alias{GradientDescent}
\title{Gradient Descent Script}
\usage{
GradientDescent(
  NumberOfEnzymes,
  InitialGuess,
  NumberOfSteps,
  GDdtMultiple,
  GDdtCutoff
)
}
\arguments{
\item{NumberOfEnzymes}{the integer number of enzymes used in the pathway}

\item{InitialGuess}{a vector of inital guess values}

\item{NumberOfSteps}{an integer}

\item{GDdtMultiple}{a float number}
}
\value{
a vector and the number of steps to get there
}
\description{
Function that takes the number of enzymes allowed to vary,
initial guess for the concentration of each of the
enzymes (inputted as a vector), number of steps to be performed,
and the multiple used to determine the step size of gradient descent.
Recommended that the multiple be below 1, however the algorithm will
adjust the step size after each iteration.
}
