% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MonteCarlo.R
\name{MC}
\alias{MC}
\title{MonteCarlo}
\usage{
MC(
  NumberOfEnzymes,
  TotalEnzyme,
  NumberOfRuns,
  ThresholdGuess,
  GDPrecision,
  GDNumberOfSteps
)
}
\value{
starting points for gradient descent algorithm
}
\description{
Function that generates the starting points to use for gradient descent,
only allowing values that perform better than an initial guess.
The MCGD algorithm is a combination of the MC and the GradientDescent functions.
}
