% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCGD.R
\name{MCGD}
\alias{MCGD}
\title{Gradient Descent using Monte Carlo sampling of starting points}
\usage{
MCGD(
  NumberOfEnzymes,
  TotalEnzyme,
  NumberOfRuns,
  ThresholdGuess,
  GDPrecision,
  GDNumberOfSteps,
  GDdtCutoff
)
}
\arguments{
\item{NumberOfEnzymes}{integer number of enzymes used in pathway}

\item{TotalEnzyme}{total enzyme}

\item{NumberOfRuns}{integer number of runs}

\item{GDPrecision}{a float number}

\item{GDNumberOfSteps}{number of steps taken}

\item{ThreshholdGuess}{a vector}
}
\value{
SolutionVector
}
\description{
This function optimizes enzyme ratios for a cell-free system.
}
