% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RateFunction1.R
\name{RateFunction1}
\alias{RateFunction1}
\title{RateFunction1}
\usage{
RateFunction1(
  EnzymeConcentration,
  TurnoverRate,
  SubstrateConcentration,
  MichaelisConstant,
  CompetitiveInhibitorConcentrations,
  CompetitiveInhibitionConstants,
  NonCompetitiveInhibitorConcentrations,
  NonCompetitiveInhibitionConstants
)
}
\arguments{
\item{EnzymeConcentration}{/}

\item{TurnoverRate}{/}

\item{SubstrateConcentration}{/}

\item{MichaelisConstant}{/}

\item{CompetitiveInhibitorConcentrations}{/}

\item{NonCompetitiveInhibitorConcentrations}{/}

\item{NonCompetitiveInhibitionConstants}{/}

\item{CompetitiveInhibitorConstants}{/}
}
\value{
Rate
}
\description{
RateFunction1
}
