% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RateFunction2.R
\name{RateFunction2}
\alias{RateFunction2}
\title{RateFunction2}
\usage{
RateFunction2(
  EnzymeConcentration,
  TurnoverRate,
  SubstrateConcentrations,
  MichaelisConstants,
  Substrate1CompetitiveInhibitorConcentrations,
  Substrate1CompetitiveInhibitionConstants,
  Substrate2CompetitiveInhibitorConcentrations,
  Substrate2CompetitiveInhibitionConstants,
  NonCompetitiveInhibitorConcentrations,
  NonCompetitiveInhibitionConstants
)
}
\arguments{
\item{EnzymeConcentration}{/}

\item{TurnoverRate}{/}

\item{SubstrateConcentrations}{/}

\item{MichaelisConstants}{/}

\item{Substrate1CompetitiveInhibitorConcentrations}{/}

\item{Substrate1CompetitiveInhibitionConstants}{/}

\item{Substrate2CompetitiveInhibitorConcentrations}{/}

\item{Substrate2CompetitiveInhibitionConstants}{/}

\item{NonCompetitiveInhibitorConcentrations}{/}

\item{NonCompetitiveInhibitionConstants}{/}
}
\description{
RateFunction2
}
