% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_function.R
\name{test_function}
\alias{test_function}
\title{Test function squares a number x}
\usage{
test_function(x)
}
\arguments{
\item{x}{input a value x}
}
\value{
x^2
}
\description{
This is a test function that takes an input x and prints its square.
}
