% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MonteCarloGradientDescent.R
\name{MCGD}
\alias{MCGD}
\title{Gradient Descent using Monte Carlo sampling of starting points}
\usage{
MCGD(
  NumberOfEnzymes,
  TotalEnzyme,
  NumberOfRuns,
  ThresholdGuess,
  GDPrecision,
  GDNumberOfSteps
)
}
\arguments{
\item{NumberOfEnzymes}{integer number of enzymes used in pathway}

\item{TotalEnzyme}{total enzyme}

\item{NumberOfRuns}{integer number of runs}

\item{GDPrecision}{a float number}

\item{GDNumberOfSteps}{number of steps taken}

\item{ThreshholdGuess}{a vector}
}
\value{
SolutionVector
}
\description{
Function that takes the same inputs as GradientDescent
as well as a total amount of enzyme to be used as
the constraint in the optimization process,
an initial guess for a starting point for gradient descent,
and the number of starting points to use.
The function incorporates Monte Carlo sampling of starting points,
and only accepts starting points that have lower values
than the value fo the initial guess value.
A default guess for the starting point can be
the enzymes expressed in a 1:1:1… ratio.
}
